+++ R/fiery-package.R
@@ -11,5 +11,6 @@
 
 ## usethis namespace: start
 #' @import rlang
+#' @importFrom lifecycle deprecated
 ## usethis namespace: end
 NULL